feat(bldr, cli, conventional-commits, core, env, fs, git, group, idx, json, num, oak, paka, resource, syn, test, tex, tree)!, feat(color, doc-inject, html, http, log, mod, name, npm-registry, release, sch, semver, url, ware), fix(flo, github, kitz): add @kitz/release package for commit-driven releases
Centralize commit-driven release planning, previewing, and application under @kitz/release.

- add the release package and move shared PR preview behavior onto the release-owned path
- move PR comment ownership into @kitz/github, including marker-aware preview comment lookup and updates
- simplify the PR workflow to call release-owned preview logic directly instead of shell glue
- remove obsolete guard-style tests/scripts and the old invalid PR body enforcement path
- align kitz hooks with the staged-snapshot pre-commit model from the reference repo

This lands the release epic as one cohesive squash commit spanning the release package, workflow, preview, and hook updates.

+++ packages/num/src/even/even.ts
@@ -42,7 +42,7 @@ export type Even = number & Brand.Brand<'Even'>
  * is('4') // false (not a number)
  */
 export const is = (value: unknown): value is Even & Int => {
-  return IntNs.is(value) && (value as Int) % 2 === 0
+  return IntNs.is(value) && value % 2 === 0
 }
 
 /**
